fix: mypy errors in format_size, request, handshake_handler, internal/network

+++ veltix/utils/format_size.py
@@ -1,4 +1,4 @@
-def format_bytes(size: int) -> str:
+def format_bytes(size: float) -> str:
     """
     Format a byte count into a human-readable string.
